Creates a draggable helper view for the analysis
