+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/composite-layer-analysis-view.js
@@ -1,10 +1,7 @@
 var template = require('./composite-layer-analysis-view.tpl');
-var LayerAnalysisHelper = require('../layer-analysis-view-helper');
+var LayerAnalysisDraggableHelperView = require('../layer-analysis-draggable-helper-view');
 
 require('jquery-ui/draggable');
-
-var DRAGGABLE_SCOPE = 'analysis';
-var REVERT_DURATION = 100;
 
 /**
  * View for an analysis node which have two source nodes as input.
@@ -42,41 +39,18 @@ module.exports = cdb.core.View.extend({
     this.addView(view);
     this.$('.js-secondary-source').append(view.render().el);
 
-    this._initDraggable();
+    this.draggableHelperView = new LayerAnalysisDraggableHelperView({
+      el: this.el
+    });
+
+    this.addView(this.draggableHelperView);
+    this.draggableHelperView.bind('dropped', this._onDropped, this);
 
     return this;
   },
 
-  _initDraggable: function () {
-    this.$el.draggable({
-      revert: true,
-      scope: DRAGGABLE_SCOPE,
-      revertDuration: REVERT_DURATION,
-      stop: this._onDraggableStop,
-      helper: this._createHelper
-    });
-  },
-
-  _createHelper: function () {
-    this._layerAnalysisHelper = new LayerAnalysisHelper();
-    return this._layerAnalysisHelper.createHelper(this);
-  },
-
-  _onDraggableStop: function (e, ui) {
-    if (ui.helper.data('dropped')) {
-      console.log(this.model.toJSON());
-    }
-  },
-
-  _destroyDraggable: function () {
-    if (this.$el.data('ui-draggable')) {
-      this.$el.draggable('destroy');
-    }
-  },
-
-  clean: function () {
-    this._destroyDraggable();
-    cdb.core.View.prototype.clean.apply(this);
+  _onDropped: function () {
+    console.log(this.model.toJSON()); // TODO: replace with actual layer generation
   }
 
 });
