+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/ref-layer-analysis-view.js
@@ -1,11 +1,5 @@
 var template = require('./ref-layer-analysis-view.tpl');
-var _ = require('underscore');
-var LayerAnalysisHelper = require('../layer-analysis-view-helper');
-
-require('jquery-ui');
-
-var DRAGGABLE_SCOPE = 'analysis';
-var REVERT_DURATION = 100;
+var LayerAnalysisDraggableHelperView = require('../layer-analysis-draggable-helper-view');
 
 /**
  * Reference to another layer.
@@ -18,7 +12,6 @@ module.exports = cdb.core.View.extend({
   className: 'js-analysis',
 
   initialize: function (opts) {
-    _.bindAll(this, '_createHelper', '_onDraggableStop');
     if (!opts.layerDefinitionModel) throw new Error('layerDefinitionModel is required');
 
     this._layerDefinitionModel = opts.layerDefinitionModel;
@@ -31,41 +24,18 @@ module.exports = cdb.core.View.extend({
       title: this.model.getName()
     }));
 
-    this._initDraggable();
+    this.draggableHelperView = new LayerAnalysisDraggableHelperView({
+      el: this.el
+    });
+
+    this.addView(this.draggableHelperView);
+    this.draggableHelperView.bind('dropped', this._onDropped, this);
 
     return this;
   },
 
-  _initDraggable: function () {
-    this.$el.draggable({
-      revert: true,
-      scope: DRAGGABLE_SCOPE,
-      revertDuration: REVERT_DURATION,
-      stop: this._onDraggableStop,
-      helper: this._createHelper
-    });
-  },
-
-  _createHelper: function () {
-    this._layerAnalysisHelper = new LayerAnalysisHelper();
-    return this._layerAnalysisHelper.createHelper(this);
-  },
-
-  _onDraggableStop: function (e, ui) {
-    if (ui.helper.data('dropped')) {
-      console.log(this.model.toJSON());
-    }
-  },
-
-  _destroyDraggable: function () {
-    if (this.$el.data('ui-draggable')) {
-      this.$el.draggable('destroy');
-    }
-  },
-
-  clean: function () {
-    this._destroyDraggable();
-    cdb.core.View.prototype.clean.apply(this);
+  _onDropped: function () {
+    console.log(this.model.toJSON()); // TODO: replace with actual layer generation
   }
 
 });
