+++ lib/assets/test/spec/cartodb3/editor/layers/analysis-view/default-layers-analysis-view.spec.js
@@ -32,15 +32,4 @@ describe('editor/layers/analysis-views/default-layer-analysis-view', function ()
     expect(this.view.$el.text()).toContain('a3');
   });
 
-  describe('draggable', function () {
-    beforeEach(function () {
-      spyOn(this.view, '_initDraggable').and.callThrough();
-      this.view.render();
-    });
-
-    it('should be initialized when view is rendered', function () {
-      expect(this.view._initDraggable).toHaveBeenCalled();
-      expect(this.view.$el.data('ui-draggable')).not.toBeUndefined();
-    });
-  });
 });
